wireguard-control: use one Key implementation for all platforms
